Refactor UI components folder structure
Moved AddToCartButton and FlavourPickerButton components and their styles into dedicated subfolders for better organization. Updated imports throughout the codebase to reflect new paths. Removed old IngredientsDropdown and its styles, as well as unused files.
