+++ src/components/PresetTextures.jsx
@@ -3,7 +3,7 @@
 import React from 'react'
 import { CONFIG_OPTIONS } from '../config/modelConfig'
 import { FLAVOR_CONFIG } from '../config/flavorConfig'
-import { FlavourPickerButton } from './ui/FlavourPickerButton'
+import { FlavourPickerButton } from './ui/FlavourPickerButton/FlavourPickerButton'
 import styles from '../styles/configurator.module.css'
 
 const PresetTextures = ({ config, setConfig }) => {
