+++ src/components/SugarFreeTextures.jsx
@@ -3,25 +3,22 @@
 import React from 'react'
 import { CONFIG_OPTIONS } from '../config/modelConfig'
 import { FLAVOR_CONFIG } from '../config/flavorConfig'
-import { FlavourPickerButton } from './ui/FlavourPickerButton'
+import { FlavourPickerButton } from './ui/FlavourPickerButton/FlavourPickerButton'
 import styles from '../styles/configurator.module.css'
 
 const SugarFreeTextures = ({ config, setConfig }) => {
 	const handleSugarFreeToggle = () => {
 		if (!config.textureUrl) return
 
-		// Hitta nuvarande smak-nyckel
 		const currentFlavorKey = Object.keys(CONFIG_OPTIONS.textures).find(
 			(key) => config.textureUrl === CONFIG_OPTIONS.textures[key]
 		)
 
 		if (!currentFlavorKey) return
 
-		// Kolla om vi är på sugar free eller regular
 		const isSugarFree = currentFlavorKey.includes('SugarFree')
 
 		if (isSugarFree) {
-			// Byt till regular version
 			const regularKey = currentFlavorKey.replace('SugarFree', '')
 			if (CONFIG_OPTIONS.textures[regularKey]) {
 				setConfig((prev) => ({
@@ -30,7 +27,6 @@ const SugarFreeTextures = ({ config, setConfig }) => {
 				}))
 			}
 		} else {
-			// Byt till sugar free version
 			const sugarFreeKey = currentFlavorKey + 'SugarFree'
 			if (CONFIG_OPTIONS.textures[sugarFreeKey]) {
 				setConfig((prev) => ({
@@ -54,7 +50,6 @@ const SugarFreeTextures = ({ config, setConfig }) => {
 		return currentKey?.includes('SugarFree') || false
 	}
 
-	// Vit knapp-konfiguration
 	const whiteButtonConfig = {
 		colors: {
 			primary: '#ffffff',
